requirements.txt
Django==4.2.1
django_environ==0.10.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2